% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Gathers useful statistics about the underlying data (mean, median, sd, counts, frequencies etc.).}
\usage{
describe(data, group = NULL)
}
\arguments{
\item{data}{The data to summarize.}

\item{group}{A grouping variable or list, if desired. This can be a character (column name), number (column index), or list of equal length to the variable in data.}
}
\value{
A data frame or list containing the results.
}
\description{
This is a slightly extended version of \code{\link[=summary]{summary()}}, combined with the functionality of \code{\link[=aggregate]{aggregate()}} - but human readable.
Factors, logicals and strings are reported up to a total of 15 different options, after which other levels are truncated in ascending order.
}
\examples{
data = data.frame(weight = rnorm(100, mean=70, sd=15),
                  gender = c(rep("male", 50), rep("female", 50)))
describe(data, "gender")
describe(data$weight, data$gender)
}
