% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.table.R
\name{freq.table}
\alias{freq.table}
\title{Frequency table with proportions}
\usage{
freq.table(
  ...,
  is.list = F,
  is.table = F,
  row.prop = T,
  col.prop = T,
  num.fmt = ".2f"
)
}
\arguments{
\item{...}{Elements to crosstabulate. This is analogous to the structure of table().}

\item{is.list}{Specifies if the supplied arguments in ... are a list and should be interpreted as a list of tables to be constructed.}

\item{is.table}{Specifies if the supplied arguments in ... are a tibble, and should have the first column interpreted as row names.}

\item{row.prop}{Specifies whether row proportions should be shown.}

\item{col.prop}{Specifies whether column proportions should be shown.}

\item{num.fmt}{Specifies the display style for the proportions in \code{\link[=sprintf]{sprintf()}}-format. Default is .2f.}
}
\value{
A formatted matrix containing either 2 columns and n rows for a one-dimensional table, or m\emph{2+1 columns and n}2+1 rows for a two-dimensional table.
}
\description{
This function combines \code{\link[=table]{table()}} and \code{\link[=proportions]{proportions()}} to output a human-readable frequency table. \code{freq.table(x,y)} is essentially a shorthand for calling \code{table(x,y)} and \code{proportions(table(x,y))} consecutively.
}
\examples{
dataset = c(rep(1, 15), rep(2, 5))
freq.table(dataset)
}
