% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemeenten.ah}
\alias{gemeenten.ah}
\title{List of municipalities in the Achterhoek region. This is a subset of \code{gemeenten.nog}.}
\format{
An object of class \code{character} of length 8.
}
\usage{
gemeenten.ah
}
\description{
List of municipalities in the Achterhoek region. This is a subset of \code{gemeenten.nog}.
}
\keyword{datasets}
