% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemeenten.nog}
\alias{gemeenten.nog}
\title{List of municipalities in the NOG region.}
\format{
An object of class \code{character} of length 22.
}
\usage{
gemeenten.nog
}
\description{
List of municipalities in the NOG region.
}
\keyword{datasets}
