% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemeenten.nv}
\alias{gemeenten.nv}
\title{List of municipalities in the Noord-Veluwe region. This is a subset of \code{gemeenten.nog}.}
\format{
An object of class \code{character} of length 6.
}
\usage{
gemeenten.nv
}
\description{
List of municipalities in the Noord-Veluwe region. This is a subset of \code{gemeenten.nog}.
}
\keyword{datasets}
