% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nog_colors}
\alias{nog_colors}
\title{Colors used in the NOG coloring scheme.}
\format{
An object of class \code{character} of length 5.
}
\usage{
nog_colors
}
\description{
Colors used in the NOG coloring scheme.
}
\keyword{datasets}
