% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nog_palette}
\alias{nog_palette}
\title{Colors used for gradients in the NOG coloring scheme.}
\format{
An object of class \code{character} of length 5.
}
\usage{
nog_palette
}
\description{
Colors used for gradients in the NOG coloring scheme.
}
\keyword{datasets}
