% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printf.R
\name{printf}
\alias{printf}
\title{Shorthand matching C-style printf() functionality.}
\usage{
printf(...)
}
\arguments{
\item{...}{A character vector, an object, or any combination of these. If an initial character vector is supplied, it is parsed using \code{\link[=sprintf]{sprintf()}}.}
}
\value{
N/A
}
\description{
Shorthand matching C-style printf() functionality.
}
\examples{
printf("\%d \%s", 1, "abc")
}
\seealso{
\code{\link[=sprintf]{sprintf()}}
}
